mutual_funds_api/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
idna==2.10
mftool==2.1
pkg-resources==0.0.0
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
